moved rating answers to separate file
