+++ src/app/features/historical-data/demo-historical-data-generator.ts
@@ -10,29 +10,6 @@ import {
 import { faker } from "../../core/demo-data/faker";
 import { ENTITY_CONFIG_PREFIX } from "../../core/entity/entity";
 
-export const ratingAnswers = [
-  {
-    id: "notTrueAtAll",
-    label: "not true at all",
-  },
-  {
-    id: "rarelyTrue",
-    label: "rarely true",
-  },
-  {
-    id: "usuallyTrue",
-    label: "usually true",
-  },
-  {
-    id: "absolutelyTrue",
-    label: "absolutelyTrue",
-  },
-  {
-    id: "noAnswerPossible",
-    label: "no answer possible",
-  },
-];
-
 export class DemoHistoricalDataConfig {
   minCountAttributes: number;
   maxCountAttributes: number;
