+++ src/app/features/historical-data/rating-answers.ts
@@ -1,0 +1,22 @@
+export const ratingAnswers = [
+  {
+    id: "notTrueAtAll",
+    label: "not true at all",
+  },
+  {
+    id: "rarelyTrue",
+    label: "rarely true",
+  },
+  {
+    id: "usuallyTrue",
+    label: "usually true",
+  },
+  {
+    id: "absolutelyTrue",
+    label: "absolutelyTrue",
+  },
+  {
+    id: "noAnswerPossible",
+    label: "no answer possible",
+  },
+];
